[HER-21] Break style and extract some components
