+++ src/about/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { People } from 'about/persons/mapping'
+import { People } from 'about/persons/people'
 import { Title } from 'about/title'
 
 export const About = () => <>
