+++ src/about/person/description/style.js
@@ -2,7 +2,7 @@ import { makeStyles } from '@material-ui/core'
 
 export const useStyles = makeStyles(() => ({
   root: {
-    margin: '0 auto',
-    maxWidth: '1700px',
+    fontStyle: 'italic',
+    lineHeight: 1.7,
   },
 }))
