+++ src/about/person/name/style.js
@@ -1,0 +1,3 @@
+import { makeStyles } from '@material-ui/core'
+
+export const useStyles = makeStyles(() => ({ root: { color: '#89CFF0' } }))
