+++ src/about/person/photo/style.js
@@ -2,7 +2,8 @@ import { makeStyles } from '@material-ui/core'
 
 export const useStyles = makeStyles(() => ({
   root: {
-    margin: '0 auto',
-    maxWidth: '1700px',
+    maxWidth: '30vw',
+    maxHeight: '30vh',
+    paddingBottom: 64,
   },
 }))
