+++ src/about/persons/people.jsx
