+++ src/index.jsx
@@ -1,8 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 
-import 'index.css'
-
 import { App } from 'app'
 import 'index.css'
